Add default shortcuts for window tiling
The Super+Left, Right shortcuts are defaults in Windows, the Up, Down complete
the tiling cycle and for corner tiling we can use the Numpad.
It may be nicer to use KP_Left instead of just Left to have everything on the
Numpad, but this would be less discoverable for people coming from Windows.