Additional open-terminal-here custom action for Thunar for files
The current "open terminal here" custom action only works on directories.
When using the list or the detail view in Thunar it's hard for the user
to open a terminal in the current directory without enabling the custom
action on files too.
As the syntax for exo-open changes from %f to %d we had to add an additional
custom action for files.